+++ lib/legion/cli/chat/tool_registry.rb
@@ -34,6 +34,7 @@ begin
   require 'legion/cli/chat/tools/detect_anomalies'
   require 'legion/cli/chat/tools/view_trends'
   require 'legion/cli/chat/tools/trigger_dream'
+  require 'legion/cli/chat/tools/generate_insights'
 rescue LoadError => e
   Legion::Logging.debug("ToolRegistry ruby_llm not available, chat tools will not be registered: #{e.message}") if defined?(Legion::Logging)
 end
@@ -74,7 +75,8 @@ module Legion
                             Tools::WorkerStatus,
                             Tools::DetectAnomalies,
                             Tools::ViewTrends,
-                            Tools::TriggerDream
+                            Tools::TriggerDream,
+                            Tools::GenerateInsights
                           ].freeze
                         else
                           [].freeze